requirements.txt
Django==1.9.8
django-registration-redux==1.4
